domain/src/test/java/com/java/althgriom/offer/TreeIsSymmetricTest.java
package com.java.althgriom.offer;

import com.java.althgriom.leetcode.entity.TreeNode;
import org.junit.Assert;
import org.junit.Test;

import static org.junit.Assert.*;

/**
 * @Description:
 * @Author: leiline
 * @CreateTime: 2020-06-30
 */
public class TreeIsSymmetricTest {

    @Test
    public void isSymmetricTest() {
        TreeNode node0 = new TreeNode(1);
        TreeNode node1 = new TreeNode(2);
        TreeNode node2 = new TreeNode(2);
        TreeNode node3 = new TreeNode(3);
        TreeNode node4 = new TreeNode(4);
        TreeNode node5 = new TreeNode(4);
        TreeNode node6 = new TreeNode(3);

        node0.left = node1;
        node0.right = node2;
        node1.left = node3;
        node1.right = node4;
        node2.left = node5;
        node2.right = node6;

        Assert.assertTrue(TreeIsSymmetric.isSymmetric(node0));
    }

    @Test
    public void isSymmetricTest2() {
        TreeNode node0 = new TreeNode(1);
        TreeNode node1 = new TreeNode(2);
        TreeNode node2 = new TreeNode(2);
        TreeNode node3 = new TreeNode(3);
        TreeNode node4 = new TreeNode(3);
        TreeNode node5 = new TreeNode(4);
        TreeNode node6 = new TreeNode(4);


        node0.left = node1;
        node0.right = node2;
        node1.left = node3;
        node2.right = node4;
        node3.left = node5;
        node4.right = node6;

        Assert.assertTrue(TreeIsSymmetric.isSymmetric(node0));
    }
}